feature: Create seperate components for TodoList and TodoItem
